requirements.txt
boto3
sklearn
pyyaml
pandas
dvc